Implementing services to pass the tests.

+++ BOAProject.Core/AppServices/Implementation/ProductService.cs
@@ -18,12 +18,28 @@ namespace BOAProject.Core.AppServices.Implementation
 
         public Product AddProduct(Product product)
         {
+            if (string.IsNullOrEmpty(product.Name))
+                throw new Exception("Name of the product is required.");
+            else if (string.IsNullOrEmpty(product.Description))
+                throw new Exception("Description of the product is required.");
+            else if (string.IsNullOrEmpty(product.Type))
+                throw new Exception("Type of the product is required.");
+            else if (product.AvailableQuantity < 0)
+                throw new Exception("Quantity of the product is wrong.");
+            else if (product.Price <= 0)
+                throw new Exception("Price of the product is wrong.");
+            else if (product.DiscountPrice < 0)
+                throw new Exception("Discount price of the product is wrong.");
+
             return _productRepo.CreateProduct(product);
         }
 
         public Product ReadProductByID(int id)
         {
-            return _productRepo.GetProductByID(id);
+            if (id <= 0)
+                throw new Exception("Minimum ID is 1.");
+            else
+                return _productRepo.GetProductByID(id);
         }
 
         public IEnumerable<Product> ReadProducts(Filter filter)
@@ -36,12 +52,28 @@ namespace BOAProject.Core.AppServices.Implementation
 
         public bool RemoveProduct(int id)
         {
-            return _productRepo.DeleteProduct(id);
+            if (id <= 0)
+                throw new Exception("Minimum ID is 1.");
+            else
+                return _productRepo.DeleteProduct(id);
         }
 
         public Product ReviseProduct(Product product)
         {
-            return _productRepo.UpdateProduct(product);
+            if (string.IsNullOrEmpty(product.Name))
+                throw new Exception("Name of the product is required.");
+            else if (string.IsNullOrEmpty(product.Description))
+                throw new Exception("Description of the product is required.");
+            else if (string.IsNullOrEmpty(product.Type))
+                throw new Exception("Type of the product is required.");
+            else if (product.AvailableQuantity < 0)
+                throw new Exception("Quantity of the product is wrong.");
+            else if (product.Price <= 0)
+                throw new Exception("Price of the product is wrong.");
+            else if (product.DiscountPrice < 0)
+                throw new Exception("Discount price of the product is wrong.");
+            else
+                return _productRepo.UpdateProduct(product);
         }
     }
 }
